# app/svc/docs/swagger.yaml
basePath: /
definitions:
...
  handler.Entry:
    properties:
      body:
        type: string
      campaignIds:
        items:
          type: integer
        type: array
      category:
        type: string
      createdAt:
        type: integer
      id:
        type: string
      linkedEntries:
        items:
          type: integer
        type: array
      ownerUserId:
        type: string
      quickViewData:
        items:
          type: integer
        type: array
      tags:
        items:
          type: integer
        type: array
      title:
        type: string
      updatedAt:
        type: integer
      visibility:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.OkResponse:
    properties:
      ok:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: 'Backend-for-frontend: session management, campaigns, entries, annotations.'
  title: Quillit BFF API
  version: "1.0"
paths:
...
  /api/entries:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Entry'
            type: array
      summary: List entries
      tags:
      - entries
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Entry'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create entry
      tags:
      - entries
  /api/entries/{id}:
    delete:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
      summary: Delete entry
      tags:
      - entries
    get:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Entry'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get entry
      tags:
      - entries
    patch:
      consumes:
      - application/json
      description: Partial update — send only the fields to change.
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (partial Entry)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.Entry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Entry'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update entry
      tags:
      - entries
  /api/entries/{id}/images:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Upload image attachment for an entry
      tags:
      - entries
...
  /api/member/folders/{id}/entries/{entryId}:
    delete:
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: Entry ID
        in: path
        name: entryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Remove entry from folder
      tags:
      - member
...
swagger: "2.0"